+++ kotlin-rpc-framework/src/main/kotlin/com/xiao/rpc/CloseableResource.kt
@@ -5,7 +5,7 @@ package com.xiao.rpc
  * @author lix wang
  */
 interface CloseableResource {
-    fun tryClose(keepAliveMills: Int): Boolean
+    fun tryClose(keepAliveMills: Long): Boolean
 
     fun tryUse(): Boolean
